sql module, tests passing
