+++ qtextasdata/__init__.py
@@ -1,0 +1,1 @@
+q_version = '3.1.6'
